warningつぶしと修正

+++ Game/Enemy.cpp
@@ -78,8 +78,9 @@ void Enemy::Draw()
 	DrawString(520, 30, _name, GetColor(255, 64, 64));
 	DrawFormatString(520, 50, GetColor(255, 64, 64), "HP: %d", _hp);
 
-	DrawBox(Transform.Position.X - Transform.Size.W / 2, Transform.Position.Y - Transform.Size.H / 2,
-		Transform.Position.X + Transform.Size.W / 2, Transform.Position.Y + Transform.Size.H / 2,
+	DrawBox(
+		(int)(Transform.Position.X - Transform.Size.W / 2), (int)(Transform.Position.Y - Transform.Size.H / 2),
+		(int)(Transform.Position.X + Transform.Size.W / 2), (int)(Transform.Position.Y + Transform.Size.H / 2),
 		GetColor(255, 64, 64), TRUE);
 }
