+++ Game/MainGameState.cpp
@@ -26,7 +26,7 @@ void MainGameState::Enter()
 	for (int i = 0; i < MAX_ENEMY; ++i)
 	{
 		_enemy[i] = std::make_shared<Enemy>();
-		_enemy[i]->Transform.Position = Vector2(280 + mt()%360, 350 - mt() % 150);
+		_enemy[i]->Transform.Position = Vector2((float)(280 + mt()%360), (float)(350 - (int)mt() % 150));
 		TaskManager::Register(_enemy[i]);
 	}
 }
